requirements.txt
aniso8601==7.0.0
annotated-types==0.6.0
anyio==4.3.0
apispec==6.6.1
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
cheroot==10.0.1
click==8.1.7
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
Flask==3.0.3
Flask-GraphQL==2.0.1
flask-smorest==0.44.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.1.1
grpcio==1.64.0
grpcio-reflection==1.64.0
grpcio-tools==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.functools==4.0.1
Jinja2==3.1.4
json5==0.9.25
jsonrpcserver==5.0.9
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
libcst==1.3.1
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
more-itertools==10.2.0
orjson==3.10.3
OSlash==0.6.3
packaging==24.0
platformdirs==4.2.1
promise==2.3
protobuf==5.26.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyodata==1.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
Rx==1.6.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
spyne==2.14.0
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
webargs==8.4.0
websockets==12.0
Werkzeug==3.0.3
WsgiDAV==4.3.3
wsproto==1.2.0
zeep==4.2.1
zipp==3.18.1
